Refactor some utils, moved some utils in post.views
